Add generic IffID3 class
